% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdf.R
\name{is_pdf}
\alias{is_pdf}
\title{Test if a file is a pdf}
\usage{
is_pdf(f)
}
\arguments{
\item{f}{Path to a file to test}
}
\value{
TRUE if pdftools::pdf_info thinks it's a PDF, else FALSE
}
\description{
Test if a file is a pdf
}
\details{
All PDFs should have an dictionary that contains information about
the key metadata about the document. This function uses
\link[pdftools]{pdf_info} to check if the downloaded file is, in fact, a PDF
because too often we get, e.g., an HTML status 404 download instead of the
desired PDF.
}

