% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdf.R
\name{make_pdf_dest}
\alias{make_pdf_dest}
\title{Construct a path to download a PDF}
\usage{
make_pdf_dest(url, subd = "")
}
\arguments{
\item{url}{The URL from which to download a PDF document}

\item{subd}{The subdirectory in which the download will be written}
}
\value{
The file path where the download will be written
}
\description{
Construct a path to download a PDF
}
\details{
Creates the directory to which the PDF will be downloaded if it
doesn't yet exist and returns the download path. The filename of the download
is based on the URL, but spaces " " are replaced with underscores "_", and
the ".pdf" prefix is appended if not present.
}

